+++ src/ArduinoJson/Serialization/Writers/StaticStringWriter.hpp
@@ -8,18 +8,14 @@
 
 namespace ARDUINOJSON_NAMESPACE {
 
-// A Print implementation that allows to write in a char[]
 class StaticStringWriter {
  public:
-  StaticStringWriter(char *buf, size_t size) : end(buf + size - 1), p(buf) {
-    *p = '\0';
-  }
+  StaticStringWriter(char *buf, size_t size) : end(buf + size), p(buf) {}
 
   size_t write(uint8_t c) {
     if (p >= end)
       return 0;
     *p++ = static_cast<char>(c);
-    *p = '\0';
     return 1;
   }
 
@@ -29,7 +25,6 @@ class StaticStringWriter {
       *p++ = static_cast<char>(*s++);
       n--;
     }
-    *p = '\0';
     return size_t(p - begin);
   }